chore:banners - disable hacks post banner
This disables the Hacks blog post banner.
I have also done some refactoring and added documentation on how banners work.

fix #5528
