+++ client/src/app.tsx
@@ -21,7 +21,8 @@ import { AppSettings } from "./app-settings";
 import { docCategory } from "./utils";
 import { Contribute } from "./community";
 import { ContributorSpotlight } from "./contributor-spotlight";
-import { Banner } from "./banners";
+
+import { Banner, hasActiveBanners } from "./banners";
 
 const AllFlaws = React.lazy(() => import("./flaws"));
 const Translations = React.lazy(() => import("./translations"));
@@ -45,13 +46,7 @@ function Layout({ pageType, children }) {
   return (
     <>
       <A11yNav />
-      {/* Commented out for now. Kept as a record/reminder of how we implement
-       banners. As of May 27, 2021 we don't have any banners to show. At all.
-
-       Note, if you do uncomment banners again (because there's one to possible
-       display), remember to go to
-       */}
-      {!isServer && <Banner />}
+      {!isServer && hasActiveBanners && <Banner />}
       <div className={`page-wrapper  ${category || ""} ${pageType}`}>
         <TopNavigation />
         {children}
