+++ client/src/banners/active-banner.tsx
@@ -2,10 +2,7 @@ import * as React from "react";
 
 import { ReactComponent as CloseIcon } from "@mdn/dinocons/general/close.svg";
 import { useGA } from "../ga-context";
-// import { COMMON_SURVEY_ID } from "./ids";
 import { REDESIGN_ANNOUNCEMENT } from "./ids";
-
-// const CATEGORY_LEARNING_SURVEY = "learning web development";
 
 // The <Banner> component displays a simple call-to-action banner at
 // the bottom of the window. The following props allow it to be customized.
@@ -64,47 +61,21 @@ function Banner(props: BannerProps) {
   );
 }
 
-// function CommonSurveyBanner({ onDismissed }: { onDismissed: () => void }) {
-//   const ga = useGA();
-
-//   return (
-//     <Banner
-//       id={COMMON_SURVEY_ID}
-//       title={"Learning web development survey"}
-//       copy={
-//         "Help us understand how to make MDN better for beginners (5 minute survey)"
-//       }
-//       cta={"Take the survey"}
-//       url="https://www.surveygizmo.com/s3/6175365/59cfad9c04cf"
-//       newWindow
-//       onDismissed={onDismissed}
-//       onCTAClick={() => {
-//         ga("send", {
-//           hitType: "event",
-//           eventCategory: CATEGORY_LEARNING_SURVEY,
-//           eventAction: "CTA clicked",
-//           eventLabel: "banner",
-//         });
-//       }}
-//     />
-//   );
-// }
+function SendToGA(eventCategory: string) {
+  const ga = useGA();
+  ga("send", {
+    hitType: "event",
+    eventCategory: eventCategory,
+    eventAction: "CTA clicked",
+    eventLabel: "banner",
+  });
+}
 
 function RedesignAnnouncementBanner({
   onDismissed,
 }: {
   onDismissed: () => void;
 }) {
-  const ga = useGA();
-  const onCTAClick = () => {
-    ga("send", {
-      hitType: "event",
-      eventCategory: REDESIGN_ANNOUNCEMENT,
-      eventAction: "CTA clicked",
-      eventLabel: "banner",
-    });
-  };
-
   return (
     <Banner id={REDESIGN_ANNOUNCEMENT} onDismissed={onDismissed}>
       <p className="mdn-cta-copy">
@@ -113,7 +84,7 @@ function RedesignAnnouncementBanner({
           href="https://hacks.mozilla.org/2022/02/a-new-year-a-new-mdn/"
           target="_blank"
           rel="noopener noreferrer"
-          onClick={onCTAClick}
+          onClick={() => SendToGA(REDESIGN_ANNOUNCEMENT)}
         >
           Learn more
         </a>{" "}
@@ -125,7 +96,7 @@ function RedesignAnnouncementBanner({
 
 // The reason we're not just exporting each individual banner is because to
 // be able to lazy-load the contents of this file it needs to export a
-// default function. This this one function is the link between the <App>
+// default function. This one function is the link between the <App>
 // and all the individual banner components.
 export default function ActiveBanner({
   id,
@@ -137,8 +108,5 @@ export default function ActiveBanner({
   if (id === REDESIGN_ANNOUNCEMENT) {
     return <RedesignAnnouncementBanner onDismissed={onDismissed} />;
   }
-  // if (id === COMMON_SURVEY_ID) {
-  //   return <CommonSurveyBanner onDismissed={onDismissed} />;
-  // }
   throw new Error(`Unrecognized banner to display (${id})`);
 }
