+++ client/src/banners/banner-utils.ts
@@ -1,0 +1,44 @@
+// Set a localStorage key with a timestamp the specified number of
+// days into the future. When the user dismisses a banner we use this
+// to prevent the redisplay of the banner for a while.
+export function setEmbargoed(id: string, days: number) {
+  try {
+    let key = `banner.${id}.embargoed_until`;
+    localStorage.setItem(
+      key,
+      String(Date.now() + Math.round(days * 24 * 60 * 60 * 1000))
+    );
+  } catch (e) {
+    // If localStorage is not supported, then embargoes are not supported.
+  }
+}
+
+// See whether the specified id was passed to setEmbargoed() fewer than the
+// specified number of days ago. We check this before displaying a banner
+// so a user does not see a banner they recently dismissed.
+export function isEmbargoed(id: string) {
+  try {
+    let key = `banner.${id}.embargoed_until`;
+    let value = localStorage.getItem(key);
+    // If it is not set, then the banner has never been dismissed
+    if (!value) {
+      return false;
+    }
+    // The value from localStorage is a timestamp that we compare to
+    // the current time
+    if (parseInt(value) > Date.now()) {
+      // If the timestamp is in the future then the banner has been
+      // dismissed and the embargo has not yet expired.
+      return true;
+    } else {
+      // Otherwise, the banner was dismissed, but the embargo has
+      // expired and we can show it again.
+      localStorage.removeItem(key);
+      return false;
+    }
+  } catch (e) {
+    // If localStorage is not supported, then the embargo feature
+    // just won't work
+    return false;
+  }
+}
