+++ client/src/banners/ids.ts
@@ -1,7 +1,5 @@
-// export const COMMON_SURVEY_ID = "common_survey_banner";
-
+// The idea is that we'll add more like this and we might want to
+// leverage older keys to see if they got enrolled in previous versions.
 export const PLUS_IDv1 = "plus_banner_v1";
 export const PLUS_IDv2 = "plus_banner_v2";
 export const REDESIGN_ANNOUNCEMENT = "redesign_announcement";
-// The idea is that we'll add more like this and we might want to
-// leverage older keys to see if they got enrolled in previous versions.
