+++ client/src/banners/index.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
-import { CRUD_MODE } from "../constants";
 
-// import { useUserData } from "../user-context";
+import { setEmbargoed, isEmbargoed } from "./banner-utils";
+import { CRUD_MODE } from "../constants";
 
 // We may or may not load any active banner. But if there's a small chance
 // that we might, it's best practice to not have to lazy-load the CSS
@@ -10,110 +10,13 @@ import { CRUD_MODE } from "../constants";
 // <ActiveBanner>, at least the CSS will be ready.
 import "./banner.scss";
 
-// import { COMMON_SURVEY_ID } from "./ids";
 import { REDESIGN_ANNOUNCEMENT } from "./ids";
 
 const ActiveBanner = React.lazy(() => import("./active-banner"));
 
-// Set a localStorage key with a timestamp the specified number of
-// days into the future. When the user dismisses a banner we use this
-// to prevent the redisplay of the banner for a while.
-function setEmbargoed(id: string, days: number) {
-  try {
-    let key = `banner.${id}.embargoed_until`;
-    localStorage.setItem(
-      key,
-      String(Date.now() + Math.round(days * 24 * 60 * 60 * 1000))
-    );
-  } catch (e) {
-    // If localStorage is not supported, then embargoes are not supported.
-  }
-}
-
-// See whether the specified id was passed to setEmbargoed() fewer than the
-// specified number of days ago. We check this before displaying a banner
-// so a user does not see a banner they recently dismissed.
-function isEmbargoed(id: string) {
-  try {
-    let key = `banner.${id}.embargoed_until`;
-    let value = localStorage.getItem(key);
-    // If it is not set, then the banner has never been dismissed
-    if (!value) {
-      return false;
-    }
-    // The value from localStorage is a timestamp that we compare to
-    // the current time
-    if (parseInt(value) > Date.now()) {
-      // If the timestamp is in the future then the banner has been
-      // dismissed and the embargo has not yet expired.
-      return true;
-    } else {
-      // Otherwise, the banner was dismissed, but the embargo has
-      // expired and we can show it again.
-      localStorage.removeItem(key);
-      return false;
-    }
-  } catch (e) {
-    // If localStorage is not supported, then the embargo feature
-    // just won't work
-    return false;
-  }
-}
-/* Not used currently
-function isPathnameIncluded(id: string, pathname: string) {
-  if (id === PLUS_IDv2) {
-    return !(pathname.includes("/plus") || pathname.includes("/signin"));
-  }
-  return false;
-}
-
-function isGeoLocationIncluded(id: string, country: string) {
-  if (id === PLUS_IDv2) {
-    return country === "United States";
-  }
-  return false;
-}
-
-function isRandomlyIncluded(id: string, chancePercentage: number) {
-  try {
-    const key = `banner.${id}.chance`;
-    const previous = localStorage.getItem(key);
-    if (previous) {
-      // The dice has been thrown before
-      return JSON.parse(previous);
-    }
-    const include = Math.random() < chancePercentage / 100;
-    localStorage.setItem(key, JSON.stringify(include));
-    return include;
-  } catch (error) {
-    // If localStorage isn't working, always bail negatively.
-  }
-  return false;
-}
-*/
+export const hasActiveBanners = false;
 
 export function Banner() {
-  /* Since auth is disabled on prod we cannot rely on this.
-
-  const userData = useUserData();
-
-  // Never return true if the whoami hasn't resolved yet, anonymous or not.
-  if (!userData) {
-    return null;
-  }
-  */
-
-  // The order of the if statements is important and it's our source of
-  // truth about which banner is "more important" than the other.
-
-  // The PLUS_IDv(N) banner depends on the following logic:
-  // 0. Is the banner not disabled by an environment variable
-  // 1. Are you not on the MDN++ page or sign in/up already.
-  // 2. Are you in the United States
-  // 3. Are you part of the 10% who are randomly selected
-  // 4. Have you not dismissed it previously
-  // 5. Have you seen a different PLUS_IDvN banner before
-  // 6. Is your locale en-US?
   if (CRUD_MODE || !isEmbargoed(REDESIGN_ANNOUNCEMENT)) {
     return (
       <React.Suspense fallback={null}>
@@ -127,22 +30,5 @@ export function Banner() {
     );
   }
 
-  // THIS CODE IS LEFT COMMENTED-OUT UNTIL WE'RE CERTAIN WE'RE NOT GOING TO USE THIS.
-  // IT'S KEPT AS REMINDER HOW TO IMPLEMENT THESE.
-  // const isEnabled = (id: string) =>
-  //   !isEmbargoed(id);
-
-  // if (isEnabled(COMMON_SURVEY_ID)) {
-  //   return (
-  //     <Suspense fallback={null}>
-  //       <ActiveBanner
-  //         id={COMMON_SURVEY_ID}
-  //         onDismissed={() => {
-  //           setEmbargoed(COMMON_SURVEY_ID, 5);
-  //         }}
-  //       />
-  //     </Suspense>
-  //   );
-  // }
   return null;
 }
